+++ frontend/package.json
@@ -13,6 +13,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-scripts": "5.0.1",
+    "resume_scanner": "file:",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
