+++ frontend/src/App.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import logo from './logo.svg';
-import './App.css';
+import './styles/App.css';
+import SignUp from './components/SignUp';
+import Login from './components/Login';
 
 function App() {
   return (
